Separate provider config functions
