+++ internal/controller/mission/mission_controller.go
@@ -21,8 +21,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"reflect"
-	"strings"
 
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
@@ -35,12 +33,8 @@ import (
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
-	controllerutil "sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
-	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 	cpv1 "github.com/crossplane/crossplane/apis/pkg/v1"
-	awsv1 "github.com/upbound/provider-aws/apis/v1beta1"
-	azrv1 "github.com/upbound/provider-azure/apis/v1beta1"
 	gcpv1 "github.com/upbound/provider-gcp/apis/v1beta1"
 
 	missionv1alpha1 "github.com/holy-tech/Mission-Control-Operator/api/mission/v1alpha1"
@@ -168,131 +162,6 @@ func (r *MissionReconciler) ReconcilePackageStatus(ctx context.Context, mission 
 	}
 	mission.Status.PackageStatus[provider.Name] = ps
 	return r.Status().Update(ctx, mission)
-}
-
-func (r *MissionReconciler) ReconcileProviderConfig(ctx context.Context, pkg *missionv1alpha1.PackageConfig, mission *missionv1alpha1.Mission) error {
-	if pkg.Provider == "GCP" {
-		return r.ReconcileProviderConfigGCP(ctx, pkg, mission)
-	}
-	if pkg.Provider == "AWS" {
-		return r.ReconcileProviderConfigAWS(ctx, pkg, mission)
-	}
-	if pkg.Provider == "AZURE" {
-		return r.ReconcileProviderConfigAzure(ctx, pkg, mission)
-	}
-	message := fmt.Sprintf("Provider %s not known", pkg.Provider)
-	return errors.New(message)
-}
-
-func (r *MissionReconciler) ReconcileProviderConfigGCP(ctx context.Context, pkg *missionv1alpha1.PackageConfig, mission *missionv1alpha1.Mission) error {
-	providerName := mission.Name + "-" + strings.ToLower(pkg.Provider)
-	providerConfig := &gcpv1.ProviderConfig{}
-	expectedProviderConfig := &gcpv1.ProviderConfig{
-		ObjectMeta: v1.ObjectMeta{
-			Name: providerName,
-		},
-		Spec: gcpv1.ProviderConfigSpec{
-			ProjectID: pkg.ProjectID,
-			Credentials: gcpv1.ProviderCredentials{
-				Source: xpv1.CredentialsSourceSecret,
-				CommonCredentialSelectors: xpv1.CommonCredentialSelectors{
-					SecretRef: &xpv1.SecretKeySelector{
-						Key: pkg.Credentials.Key,
-						SecretReference: xpv1.SecretReference{
-							Name:      pkg.Credentials.Name,
-							Namespace: pkg.Credentials.Namespace,
-						},
-					},
-				},
-			},
-		},
-	}
-	if err := controllerutil.SetControllerReference(mission, expectedProviderConfig, r.Scheme); err != nil {
-		return err
-	}
-	if err := r.Get(ctx, types.NamespacedName{Name: providerName}, providerConfig); err != nil {
-		if k8serrors.IsNotFound(err) {
-			return r.Create(ctx, expectedProviderConfig)
-		}
-	} else if !reflect.DeepEqual(providerConfig, expectedProviderConfig) {
-		providerConfig.Spec = expectedProviderConfig.Spec
-		err := r.Update(ctx, providerConfig)
-		return err
-	}
-	return nil
-}
-
-func (r *MissionReconciler) ReconcileProviderConfigAWS(ctx context.Context, pkg *missionv1alpha1.PackageConfig, mission *missionv1alpha1.Mission) error {
-	providerName := mission.Name + "-" + strings.ToLower(pkg.Provider)
-	providerConfig := &awsv1.ProviderConfig{}
-	expectedProviderConfig := &awsv1.ProviderConfig{
-		ObjectMeta: v1.ObjectMeta{
-			Name: providerName,
-		},
-		Spec: awsv1.ProviderConfigSpec{
-			Credentials: awsv1.ProviderCredentials{
-				Source: xpv1.CredentialsSourceSecret,
-				CommonCredentialSelectors: xpv1.CommonCredentialSelectors{
-					SecretRef: &xpv1.SecretKeySelector{
-						Key: pkg.Credentials.Key,
-						SecretReference: xpv1.SecretReference{
-							Name:      pkg.Credentials.Name,
-							Namespace: pkg.Credentials.Namespace,
-						},
-					},
-				},
-			},
-		},
-	}
-	if err := controllerutil.SetControllerReference(mission, expectedProviderConfig, r.Scheme); err != nil {
-		return err
-	}
-	if err := r.Get(ctx, types.NamespacedName{Name: providerName}, providerConfig); err != nil {
-		if k8serrors.IsNotFound(err) {
-			return r.Create(ctx, expectedProviderConfig)
-		}
-	} else if !reflect.DeepEqual(providerConfig, expectedProviderConfig) {
-		providerConfig.Spec = expectedProviderConfig.Spec
-		err := r.Update(ctx, providerConfig)
-		return err
-	}
-	return nil
-}
-func (r *MissionReconciler) ReconcileProviderConfigAzure(ctx context.Context, pkg *missionv1alpha1.PackageConfig, mission *missionv1alpha1.Mission) error {
-	providerName := mission.Name + "-" + strings.ToLower(pkg.Provider)
-	providerConfig := &azrv1.ProviderConfig{}
-	expectedProviderConfig := &azrv1.ProviderConfig{
-		ObjectMeta: v1.ObjectMeta{
-			Name: providerName,
-		},
-		Spec: azrv1.ProviderConfigSpec{
-			Credentials: azrv1.ProviderCredentials{
-				Source: xpv1.CredentialsSourceSecret,
-				CommonCredentialSelectors: xpv1.CommonCredentialSelectors{
-					SecretRef: &xpv1.SecretKeySelector{
-						Key: pkg.Credentials.Key,
-						SecretReference: xpv1.SecretReference{
-							Name:      pkg.Credentials.Name,
-							Namespace: pkg.Credentials.Namespace,
-						},
-					},
-				},
-			},
-		},
-	}
-	if err := controllerutil.SetControllerReference(mission, expectedProviderConfig, r.Scheme); err != nil {
-		return err
-	}
-	if err := r.Get(ctx, types.NamespacedName{Name: providerName}, providerConfig); err != nil {
-		if k8serrors.IsNotFound(err) {
-			return r.Create(ctx, expectedProviderConfig)
-		}
-	} else if !reflect.DeepEqual(providerConfig, expectedProviderConfig) {
-		providerConfig.Spec = expectedProviderConfig.Spec
-		err := r.Update(ctx, providerConfig)
-		return err
-	}
-	return nil
 }
 
 func (r *MissionReconciler) SetupWithManager(mgr ctrl.Manager) error {
